+++ exports/subscriber.go
@@ -2,7 +2,6 @@ package main
 
 /*
 #include <stdlib.h>
-#include <stdint.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include "./error.h"
@@ -11,7 +10,7 @@ package main
 typedef struct {
 	bool enabled;
 	char *mode;
-	uint64_t omittedCount;
+	long long unsigned int omittedCount;
 } SubscriberInitialLoadOptions;
 
 typedef struct {
@@ -28,7 +27,7 @@ typedef struct {
 
 typedef struct {
 	void *instance;
-	uint64_t pipelineId;
+	long long unsigned int pipelineId;
 	char *eventName;
 	char *collection;
 	char *payload;
